xml_reflection: Moved into urdf_parser_py
