+++ urdf_parser_py/src/urdf_parser_py/urdf.py
@@ -1,5 +1,5 @@
-from xml_reflection.basics import *
-import xml_reflection as xmlr
+from urdf_parser_py.xml_reflection.basics import *
+import urdf_parser_py.xml_reflection as xmlr
 
 # Add a 'namespace' for names so that things don't conflict between URDF and SDF?
 # A type registry? How to scope that? Just make a 'global' type pointer?
