+++ urdf_parser_py/src/urdf_parser_py/xml_reflection/core.py
@@ -1,4 +1,4 @@
-from xml_reflection.basics import *
+from urdf_parser_py.xml_reflection.basics import *
 import sys
 import copy
