--- xml_reflection/src/xml_reflection/__init__.py
@@ -1,2 +1,0 @@
-#from xml_reflection.basics import *
-from xml_reflection.core import *
